CLN : Move docs examples into the docs tree
and literalinclude the example files in the documentation rather than
manually duplicating the information in the docs again

this commit only moves examples and removes duplicates from factory
intro

	new file:   docs/source/traitsui_user_manual/examples/default_trait_editors.py
	new file:   docs/source/traitsui_user_manual/examples/mixed_styles.py
	modified:   docs/source/traitsui_user_manual/factory_intro.rst
	deleted:    examples/tutorials/doc_examples/examples/default_trait_editors.py
	deleted:    examples/tutorials/doc_examples/examples/mixed_styles.py
